% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_wafer.R
\name{compare_wafer}
\alias{compare_wafer}
\title{Compare wafer maps}
\usage{
compare_wafer(wafer1, wafer2, output = "compare")
}
\arguments{
\item{wafer1}{thickness profile of wafer 1}

\item{wafer2}{thickness profile of wafer 2}

\item{output}{'\code{compare}': show countour of each wafer, display boxplot and radial plots,  '\code{delta}': show countour of each wafer, display delta map and radial plots,}
}
\value{
Plots for wafer map comparison
}
\description{
Takes wafer inputs in X,Y,z format and compare contour map, boxplot and radial profile
}
