% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_klaraw.R
\name{explore_klaraw}
\alias{explore_klaraw}
\title{Explore wafer profile in each slot in input files and output boxplot or summary}
\usage{
explore_klaraw(file, output = "boxplot")
}
\arguments{
\item{file}{KLA raw data file of wafer profile measurement}

\item{output}{specify output type: '\code{boxplot}': show a boxplot, '\code{summary}': show a summary}
}
\value{
A boxplot or a summary of the wafer profiles in the input file
}
\description{
A quick exploration of wafer profile can be done with KLA raw data files using this function. It outputs a boxplot or a summary of the input file.
}
