% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explore_lot_klaraw.R
\name{explore_lot_klaraw}
\alias{explore_lot_klaraw}
\title{Explore pre and post measurement of each slot in input files}
\usage{
explore_lot_klaraw(pre, post)
}
\arguments{
\item{pre}{KLA raw data file of pre-process measurement}

\item{post}{KLA raw data file of post-process measurement}
}
\value{
the pre, post and delta thickness and a table with uniformity statistics
}
\description{
A quick exploration of wafer maps can be done with KLA raw data files using this function. It compares the pre and post measurement of each slot in the file.
}
