% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_decompose.R
\name{map_decompose}
\alias{map_decompose}
\title{Decompose wafer maps}
\usage{
map_decompose(wafer, resolution = "normal", point = "off", maps = "mean",
  wlabel = "Wafer Map")
}
\arguments{
\item{wafer}{dataframe from the wafer list (output of read.kla_f5, or a self-imported file obeys the same data structure)}

\item{resolution}{takes velue of '\code{high}', '\code{normal}', '\code{low}', '\code{rapid}'}

\item{point}{'\code{on} (\code{off})': display (hide) point label,}

\item{maps}{'\code{mean}' shows the mean map only, '\code{decompose}' all the radial and residual maps, '\code{all}' includes all above three maps and radial spatial distribution}

\item{wlabel}{label of the output plot}
}
\value{
coresponding wafer map(s)
}
\description{
Given thickness profile of a wafer, this function plots wafer maps including radial and residual decomposition
}
