% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_kla_f5.R
\name{read_kla_f5}
\alias{read_kla_f5}
\title{Read from raw data file from KLA-Tencor metrology tools}
\usage{
read_kla_f5(filename)
}
\arguments{
\item{filename}{Path of the raw metrology datafile}
}
\value{
A list of wafer profiles
}
\description{
Takes raw datafile from KLA metrology machine and returns a wafer data in lists
}
