% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wafer_stat.R
\name{wafer_stat}
\alias{wafer_stat}
\title{Wafer statistics: mean, 3s and range non-uniformity (NU)}
\usage{
wafer_stat(wafer)
}
\arguments{
\item{wafer}{dataframe from the wafer list (output of read.kla_f5)}
}
\value{
Wafer statistics: mean, 3-sigma and range NU
}
\description{
Takes a dataframe from wafer list and returns wafer statistics, including mean, 3 sigma and range nonuniformity
}
